Updating required compilers' attributes
