+++ python/vimhdl/compilers/base_compiler.py
@@ -32,8 +32,15 @@ class BaseCompiler(object):
     # Shell accesses must be atomic
     _lock = Lock()
 
+    def _setName(self, value):
+        self.__builder_name__ = value
+    def _getName(self):
+        return self.__builder_name__
+
+    __name__ = abc.abstractproperty(_setName, _getName)
+
     def __init__(self, target_folder):
-        self._logger = logging.getLogger(__name__)
+        self._logger = logging.getLogger(__name__ + '.' + self.__builder_name__)
         self._target_folder = os.path.abspath(os.path.expanduser(target_folder))
 
         self.builtin_libraries = []
@@ -96,7 +103,8 @@ class BaseCompiler(object):
             self._logger.info("Forcing build of %s", str(source))
         elif source.getmtime() > cached_info['compile_time']:
             build = True
-            self._logger.info("Building %s because it's out of date", str(source))
+            self._logger.info("Building %s because it's out of date", \
+                    str(source))
 
         if build:
             # Build a set of unique flags and pass it as tuple
@@ -128,4 +136,84 @@ class BaseCompiler(object):
             rebuilds = cached_info['rebuilds']
 
         return records, rebuilds
+
+from prettytable import PrettyTable
+# TODO: Move this to a tests folder or something
+def isRecordValid(record):
+    is_valid = True
+
+    t = PrettyTable(["field", "value", "status"])
+    t.align['status'] = 'l'
+    record_info = dict(zip(record.keys(), ' '*len(record)))
+
+    # Error type check
+    if record['error_type'] in ('E', 'W'):
+        record_info['error_type'] = ''
+    else:
+        record_info['error_type'] = 'Invalid value'
+
+    # Line number check
+    try:
+        int(record['line_number'])
+        record_info['line_number'] = ''
+    except Exception as e:
+        if record['error_number'] not in ('vcom-11', 'vcom-13', 'vcom-19', 'vcom-7'):
+            record_info['line_number'] = 'NOK: %s: %s' % (e.__class__.__name__, str(e))
+            is_valid = False
+
+    # Error number check
+    if record['error_number'] is None:
+        record_info['error_number'] = 'No error number found'
+    else:
+        _errors = []
+        if record['error_number'] == ' ':
+            _errors += ['ops...']
+        for _d in record['error_number']:
+            if _d in ('[', ']', '(', ')'):
+                _errors += ['invalid char: ' + _d]
+                break
+        if _errors:
+            record_info['error_number'] = "NOK: " + ", ".join(_errors)
+            is_valid = False
+
+    # Error message check
+    _errors = []
+    if record['error_message'] == '':
+        _errors += ['empty']
+    if type(record['error_message']) is not str:
+        _errors += ['not string']
+
+    if _errors:
+        record_info['error_message'] = "NOK: " + ", ".join(_errors)
+        is_valid = False
+
+    # Filename check
+    if record['error_number'] not in ('vcom-11', 'vcom-13', 'vcom-19', 'vcom-7'):
+        _errors = []
+        if record['filename'] == '':
+            _errors += ['empty']
+        if type(record['filename']) is not str:
+            _errors += ['not string']
+        else:
+            if 'souto' in record['filename']:
+                if not os.path.isfile(record['filename']):
+                    _errors += ['file %s should exist!' % repr(record['filename'])]
+
+        if _errors:
+            record_info['filename'] = "NOK: " + ", ".join(_errors)
+            is_valid = False
+
+
+    for k, v in record.items():
+        if len(str(v)) > 80:
+            v = str(v)[:80] + '...'
+        info = record_info[k]
+        if len(str(info)) > 80:
+            info = str(info)[:80] + '...'
+        t.add_row([k, v, info])
+
+    if not is_valid:
+        print t
+
+    return is_valid
